requirements.txt
Flask
MySQL-python
openai